feat: implement InvoicesTableItem component
